Linting, formatting, etc

+++ src/asm.rs
@@ -1,3 +1,4 @@
+#[allow(dead_code)]
 unsafe extern "C" {
 	pub fn asm_syscall_hook();
 	pub fn restore_selector_trampoline();
@@ -12,9 +13,9 @@ pub extern "C" fn teardown_thread_metadata() {
 	unsafe {
 		// Disable SUD
 		let result = crate::ffi::syscall6(
-			libc::SYS_prctl as i64,
-			crate::ffi::PR_SET_SYSCALL_USER_DISPATCH as i64,
-			crate::ffi::PR_SYS_DISPATCH_OFF as i64,
+			libc::SYS_prctl,
+			i64::from(crate::ffi::PR_SET_SYSCALL_USER_DISPATCH),
+			i64::from(crate::ffi::PR_SYS_DISPATCH_OFF),
 			0,
 			0,
 			0,
@@ -25,7 +26,7 @@ pub extern "C" fn teardown_thread_metadata() {
 		// Unmap GSRelData
 		let gs_base = crate::ffi::get_gs_base();
 		let result = crate::ffi::syscall6(
-			libc::SYS_munmap as i64,
+			libc::SYS_munmap,
 			gs_base as i64,
 			4096, // Page size
 			0,
@@ -41,5 +42,5 @@ pub extern "C" fn teardown_thread_metadata() {
 pub extern "C" fn handle_clone_thread(a1: i64, a2: i64, a3: i64, a4: i64, a5: i64, a6: i64) -> i64 {
 	// In a real implementation, this would handle thread creation
 	// For now, just forward to syscall6
-	unsafe { crate::ffi::syscall6(libc::SYS_clone as i64, a1, a2, a3, a4, a5, a6) as i64 }
+	unsafe { crate::ffi::syscall6(libc::SYS_clone, a1, a2, a3, a4, a5, a6) as i64 }
 }
